requirements.txt
debugpy==1.8.17
RPi.GPIO==0.7.1
RPLCD==1.4.0
paho-mqtt==1.6.1
PyYAML==6.0.1
Jinja2==3.1.2
smbus2==0.5.0
